# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MapleBit",
        "description": "MapleBit API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "team@"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/maplebit/public/api"
        }
    ],
    "paths": {
        "/user/disconnect": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Disconnect the authenticated user account",
                "operationId": "disconnectAccount",
                "responses": {
                    "200": {
                        "description": "Indicates the account has been updated with loggedin=0"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/update": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Modify the password of the authenticated user account",
                "operationId": "updateAccount",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "password",
                                    "new_password",
                                    "new_verify_password"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "current password",
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "description": "new password",
                                        "type": "string"
                                    },
                                    "new_verify_password": {
                                        "description": "confirm new password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Indicates the account has been updated."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/me": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get basic information of the authenticated user.",
                "operationId": "getMe",
                "responses": {
                    "200": {
                        "description": "Successfully returned basic information"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/articles/list/{category?}": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Gets a paginated view of articles of optional category (omitting results in all articles)",
                "operationId": "list",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "optionally specify category to return",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "paginated list of articles"
                    }
                }
            }
        },
        "/articles/store": {
            "post": {
                "tags": [
                    "articles"
                ],
                "summary": "Store an article in the database. Requires admin or moderator role.",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "category"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "article title",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "article content",
                                        "type": "string"
                                    },
                                    "category": {
                                        "description": "article category",
                                        "type": "string"
                                    },
                                    "slug": {
                                        "description": "article slug",
                                        "type": "string"
                                    },
                                    "locked": {
                                        "description": "if article can be commented on",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article creation was successful with article data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/articles/view/{article}": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Gets specified article slug or 404",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "article slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "article title, content, category, slug, timestamps"
                    }
                }
            }
        },
        "/articles/update/{article}": {
            "put": {
                "tags": [
                    "articles"
                ],
                "summary": "Updates the specified article. Requires admin or moderator role.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "article slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "category"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "article title",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "article content",
                                        "type": "string"
                                    },
                                    "category": {
                                        "description": "article category",
                                        "type": "string"
                                    },
                                    "slug": {
                                        "description": "article slug",
                                        "type": "string"
                                    },
                                    "locked": {
                                        "description": "if article can be commented on",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article update was successful with updated article data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/articles/delete/{article}": {
            "delete": {
                "tags": [
                    "articles"
                ],
                "summary": "Deletes the specified article. Requires admin or moderator role.",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "article slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article delete was successful."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login as a user",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "user@email.com",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success with token"
                    }
                }
            }
        },
        "/auth/create": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Create a user",
                "operationId": "create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "username",
                                    "password",
                                    "password_confirm"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "user@email.com",
                                        "type": "string"
                                    },
                                    "username": {
                                        "description": "username",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string"
                                    },
                                    "password_confirm": {
                                        "description": "Password Confirmation",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success with token"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Destroys current session",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Indicates current session is destroyed"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/auth/revokeAll": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Destroys all sessions other than the one used to make this request",
                "operationId": "revokeAll",
                "responses": {
                    "200": {
                        "description": "Indicates sessions were destroyed"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/auth/devices": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Gets all authenticated devices (sessions) for user",
                "operationId": "getUserDevices",
                "responses": {
                    "200": {
                        "description": "Device identifiers, names, last used"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}